[lang] refactor ir module to factor out params and utils
