+++ lang/tests/ui/fail/01-constructor-returns.stderr
@@ -1,5 +1,5 @@
 error: constructors in ink! must have no specified return type
-  --> $DIR/01-constructor-returns.rs:16:27
+  --> $DIR/01-constructor-returns.rs:15:27
    |
-16 |         fn new(&mut self) -> Self {}
+15 |         fn new(&mut self) -> Self {}
    |                           ^^^^^^^
